Rename migrate to upgrade in code
Trello: https://trello.com/c/1KV8NmDJ (online_runtime_319)
